+++ test/helpers/access.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const fs = require('fs-extra')
+const fs = require('node:fs/promises')
 const retry = require('promise-retry')
 
 /**